requirements.txt
SpeechRecognition==3.8.1
pyttsx3==2.90
pyaudio==0.2.13
wikipedia==1.4.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
playsound==1.3.0
